Fix issue where span creation was not allowing List[Any]
This change is primarily to fix an issue where if you try to create a trace span using a List instead of a dictionary or BaseModel it errored out even though the upper function signature specified that List[Any] is allowed.

To be able to verify this with built in tests I had to update the requirements * .lock files and then the uv.lock file also got updated.

This all resulted in no longer needing or being required to limit the nox session to pydantic 1.x and it was failing because the code wants pydantic 2 so i removed the pydantic 1.x requirement from the nox file.